feat: consolidate per-step config buttons into a single Configure dialog
Addresses #894. The cell metadata editor had a separate button + dialog
per config option (IMAGE, GPU, CACHE), and reviewers on #889 and #891
pointed out this doesn't scale - each new option adds another button,
which was already about to become five with those two PRs pending.

Replace the three standalone dialogs (BaseImageDialog, GpuDialog,
CacheDialog) with a single StepConfigDialog behind one gear-icon
"Configure step" button, using tabs (Base Image / GPU / Caching)
instead of separate buttons. Tabs are wired through a `{label,
render}[]` array that Tabs/DialogContent map over, so adding a new
per-step option going forward means appending one entry to that array,
not adding another button/dialog pair or a fourth hardcoded tab index.

No backend or tag-format changes - this is a pure UI reorganization.
Updated the Playwright e2e assertion that checked for the three
separate button titles to check for the new button's aria-label.

Tested manually end-to-end in a running JupyterLab:
- Opened the Configure dialog and confirmed all three tabs render
  their original content, pre-populated correctly from existing
  limit:/cache:/image: tags.
- Confirmed Base Image, GPU, and Caching each write the correct
  notebook tag when changed (verified via live React state and, for
  GPU, a round trip to the saved .ipynb).
- Confirmed the "Reset to Default" button both clears the value and
  closes the dialog.
- Confirmed the dialog always reopens on the Base Image tab, even
  after being left on a different tab in a previous session, by
  resetting activeTab whenever the dialog opens.
- Confirmed Caching shows the correct radio immediately with no
  flash of the wrong value, since its value is derived directly from
  the enableCaching prop instead of synced into local state.
- Confirmed the dialog's aria-labelledby/aria-describedby are present
  (parity with the original GPU dialog's accessibility wiring).
